Added ui representation for checkin(booking)
TODO -add support for chekins(booking): creation, editing, simple logic
     -add support for admin system

+++ src/main/java/com/example/hotel_jpa/repositories/ICheckInRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ICheckInRepository extends JpaRepository<CheckIn, Integer> {
+public interface ICheckInRepository extends JpaRepository<CheckIn, Long> {
 
     @Query("select ch from CheckIn ch where ch.client = :client")
     CheckIn findByClient(@Param("client") Client client);
